Timeseries

+++ include/QuanSet.h
@@ -12,7 +12,7 @@ class QuanSet
         QuanSet(const QuanSet& other);
         QuanSet& operator=(const QuanSet& other);
         bool Append(const string &s, const Quan &q);
-        int Count(const string &s) {return quans.count(s);};
+        size_t Count(const string &s) {return quans.count(s);};
         Quan& operator[] (const string &s);
         std::map<string,Quan>::iterator find(const string &name) {return quans.find(name);};
         std::map<string,Quan>::iterator end() {return quans.end();};
